requirements/common.txt
click>=6.6
PyYAML>=3.11
tweepy>=3.5.0
